requirements.txt
urllib3==2.2.1
aiohttp==3.9.5
sqlalchemy==2.0.29
psycopg2_binary==2.9.9
psycopg==3.2.2
pyTelegramBotAPI==4.17.0
requests==2.32.2
langchain-community==0.3.2
langchain==0.3.3
langchain-postgres==0.0.12
langchain-ollama==0.2.0
python-telegram-bot==21.7
unstructured==0.15.12
nltk==3.9.1
lark==1.2.2
streamlit==1.38.0
pydantic==2.9.2
docx==0.2.4
scipy==1.14.1